将IBP波形区域选项移动到IBP菜单

+++ Params/IBPParam/IBPWidgets/IBPWaveRuler.h
@@ -22,13 +22,13 @@ class IBPWaveRuler : public WaveWidgetItem
 {
 public:
     virtual void paintItem(QPainter &painter);
-    void setRuler(double up, double mid, double low);
+    void setRuler(int up, int mid, int low);
 
     explicit IBPWaveRuler(IBPWaveWidget *wave);
     ~IBPWaveRuler();
 
 private:
-    double _up;             // 上标尺数值, 单位%
-    double _mid;            // 中标尺数值, 单位%
-    double _low;            // 下标尺数值, 单位%
+    int _up;             // 上标尺数值, 单位%
+    int _mid;            // 中标尺数值, 单位%
+    int _low;            // 下标尺数值, 单位%
 };
